feat: implement events overview with tabs, data table, and statistics grid
